[TEP-0091] Add Sign CMD and Verification at reconciler
This commit is part of TEP-0091, before this commit we only have signing
and verification functions but not used. This commit add signing in cmd
so users can run signing with key files or KMS key to sign Task or
Pipeline. Verification is added at reconciler after remote resolution
and local resolve is done.